+++ tests/pr_test.go
@@ -39,22 +39,6 @@ var permanentResources map[string]interface{}
 
 var sharedInfoSvc *cloudinfo.CloudInfoService
 
-var validRegions = []string{
-	"us-south",
-	"eu-de",
-	"eu-gb",
-	"jp-tok",
-	"us-east",
-	// "ca-tor",
-	// "au-syd",  Excluded regions (ca-tor, au-syd) as they are supported only in the lakehouse-enterprise-mcsp plan; this test targets enterprise plan with KMS.
-}
-
-// validMCSPRegion is a subset of validRegions that are supported for MCSP plans.
-var validMCSPRegion = []string{
-	"ca-tor",
-	"au-syd",
-}
-
 // TestMain will be run before any parallel tests, used to read data from yaml for use with tests
 func TestMain(m *testing.M) {
 	var err error
@@ -69,23 +53,6 @@ func TestMain(m *testing.M) {
 	}
 
 	os.Exit(m.Run())
-}
-
-func setupOptions(t *testing.T, prefix string, dir string) *testhelper.TestOptions {
-	options := testhelper.TestOptionsDefaultWithVars(&testhelper.TestOptions{
-		Testing:       t,
-		TerraformDir:  dir,
-		Prefix:        prefix,
-		ResourceGroup: resourceGroup,
-	})
-	options.TerraformVars = map[string]interface{}{
-		"access_tags":    permanentResources["accessTags"],
-		"region":         validMCSPRegion[common.CryptoIntn(len(validMCSPRegion))],
-		"prefix":         options.Prefix,
-		"resource_group": resourceGroup,
-		"resource_tags":  options.Tags,
-	}
-	return options
 }
 
 // Provision KMS - Key Protect to use in DA tests
@@ -131,27 +98,6 @@ func cleanupResources(t *testing.T, terraformOptions *terraform.Options, prefix 
 	}
 }
 
-func TestRunBasicExample(t *testing.T) {
-	t.Parallel()
-
-	options := setupOptions(t, "wxd-basic", basicExampleDir)
-
-	output, err := options.RunTestConsistency()
-	assert.Nil(t, err, "This should not have errored")
-	assert.NotNil(t, output, "Expected some output")
-}
-
-func TestRunAdvancedExample(t *testing.T) {
-	t.Parallel()
-
-	options := setupOptions(t, "wxd-advanced", advancedExampleDir)
-	options.TerraformVars["region"] = validRegions[common.CryptoIntn(len(validRegions))]
-
-	output, err := options.RunTestConsistency()
-	assert.Nil(t, err, "This should not have errored")
-	assert.NotNil(t, output, "Expected some output")
-}
-
 func TestRunExistingResourcesExample(t *testing.T) {
 	t.Parallel()
 
@@ -175,7 +121,7 @@ func TestRunExistingResourcesExample(t *testing.T) {
 		TerraformDir: tempTerraformDir + "/tests/existing-resources",
 		Vars: map[string]interface{}{
 			"prefix":        prefix,
-			"region":        validRegions[common.CryptoIntn(len(validRegions))],
+			"region":        "us-south",
 			"resource_tags": tags,
 			"access_tags":   permanentResources["accessTags"],
 		},
@@ -212,7 +158,7 @@ func TestRunExistingResourcesExample(t *testing.T) {
 }
 
 func setupFullyConfigurableOptions(t *testing.T, prefix string) *testschematic.TestSchematicOptions {
-	var region = validRegions[common.CryptoIntn(len(validRegions))]
+	var region = "us-south"
 	prefixKMSKey := fmt.Sprintf("%s-key", prefix)
 	prefixKMSKey += strconv.Itoa(common.CryptoIntn(1000))
 	existingTerraformOptions := setupKMSKeyProtect(t, region, prefixKMSKey)
